Set up basic outline for input limits.
Made attachHandlers function which calls check(property) when that property is inputed. Created enableButton and validations. The point is to call each check function and set a validation to true, then call enable button which only enables the sell button if all validations are true.

Next, I plan to add more limits to input fields